Renamed MapComponent files for cleaning up tests
